wire up done button
